move config files to train class, finalize minGRU model for testing
