+++ src/models/MinGRUSynthetic.py
@@ -10,7 +10,7 @@ class MinGRUSynthetic(nn.Module):
         self.embedding = nn.Embedding(vocab_size, embedding_dim)
         
         # MinGRU layers
-        self.layers = nn.ModuleList(MinGRUBlock(embedding_dim) for _ in range(num_layers))
+        self.layers = nn.ModuleList(MinGRUBlock(embedding_dim, bidirectional=bidirectional) for _ in range(num_layers))
 
         # Classifier head
         self.linear = nn.Linear(embedding_dim, num_classes)
